# Array.choose with less memory use