This mirrors the recently merged change to Array.filter

```ini

Host Process Environment Information:
BenchmarkDotNet=v0.9.8.0
OS=Microsoft Windows NT 6.2.9200.0
Processor=Intel(R) Core(TM) i7-4712HQ CPU 2.30GHz, ProcessorCount=8
Frequency=2240910 ticks, Resolution=446.2473 ns, Timer=TSC
CLR=MS.NET 4.0.30319.42000, Arch=64-bit RELEASE [RyuJIT]
GC=Concurrent Workstation
JitModules=clrjit-v4.6.1590.0

Type=CoreBenchmark  Mode=Throughput  Platform=X64  
Jit=RyuJit  GarbageCollection=Concurrent Workstation  

```
 Method | Length | Pct |        Median |     StdDev | Scaled |    Gen 0 |    Gen 1 |    Gen 2 | Bytes Allocated/Op |
------- |------- |---- |-------------- |----------- |------- |--------- |--------- |--------- |------------------- |
    **Old** | **100000** |   **0** |   **333.8228 us** |  **4.5051 us** |   **1.29** |        **-** |        **-** |   **549.04** |         **166,044.45** |
    New | 100000 |   0 |   258.9620 us |  2.5038 us |   1.00 |        - |        - |        - |             416.20 |
    **Old** | **100000** |   **2** |   **374.7043 us** |  **4.3348 us** |   **1.04** |        **-** |        **-** |   **556.34** |         **192,624.99** |
    New | 100000 |   2 |   358.8436 us |  7.1316 us |   1.00 |        - |        - |   131.93 |          66,198.04 |
    **Old** | **100000** |  **10** |   **502.1899 us** | **10.7885 us** |   **1.05** |        **-** |        **-** |   **576.07** |         **300,879.20** |
    New | 100000 |  10 |   479.0709 us |  6.4333 us |   1.00 |   285.45 |        - |   132.23 |         165,014.53 |
    **Old** | **100000** |  **25** |   **847.1055 us** | **20.8163 us** |   **1.25** |   **818.53** |        **-** |   **547.69** |         **453,030.09** |
    New | 100000 |  25 |   675.0087 us | 12.8196 us |   1.00 |   823.61 |        - |   263.88 |         331,434.90 |
    **Old** | **100000** |  **50** | **1,082.4268 us** |  **8.0474 us** |   **1.04** | **1,207.21** |        **-** |   **792.18** |         **748,076.94** |
    New | 100000 |  50 | 1,045.2314 us | 24.7514 us |   1.00 | 1,070.52 |        - |   532.97 |         562,502.02 |
    **Old** | **100000** |  **75** | **1,055.4289 us** | **18.9524 us** |   **1.02** | **1,735.12** |        **-** |   **609.26** |         **717,640.67** |
    New | 100000 |  75 | 1,030.5663 us | 23.8725 us |   1.00 | 1,734.00 |        - |   863.00 |       1,002,876.51 |
    **Old** | **100000** |  **90** | **1,373.0000 us** | **82.1311 us** |   **1.61** | **2,162.72** |        **-** |   **633.20** |         **871,658.29** |
    New | 100000 |  90 |   851.1940 us | 14.9845 us |   1.00 | 2,492.63 |        - |   981.62 |       1,223,572.50 |
    **Old** | **100000** |  **98** |   **919.6817 us** | **24.9634 us** |   **0.98** | **1,684.15** |        **-** |   **825.47** |         **989,380.17** |
    New | 100000 |  98 |   937.5664 us | 37.0001 us |   1.00 | 1,535.14 |        - |   751.91 |         901,410.26 |
    **Old** | **100000** | **100** |   **769.9352 us** | **20.6047 us** |   **0.97** | **2,185.81** | **1,037.84** | **1,052.77** |       **1,285,309.62** |
    New | 100000 | 100 |   792.3853 us |  9.6276 us |   1.00 | 2,428.47 | 1,153.57 | 1,170.37 |       1,428,041.67 |