@KevinRansom 2016-08-23
@jackmott 
Okay this looks good to me.  Did you try factoring out and sharing code with filter?