@jackmott 2016-08-23
I have not tried that yet. It may be possible.